first version of mhx file support impl.
o it can only handle S2 records (which means max. 6 bytes for the addresses) by now,
but other data records shouldn't be used by the fujitsu converter anyway i guess?
o data is only read and printed to stdout, but not really used, that will be done
in the next commit