okay, here is a programmer (=kernel.py) which communicate to the pkernel running on the board. there are some operation done to the flash, however it isn't correct (nothing happens after a reset)
note: atm you have to program the pkernel with a windows programmer.
note: use simpleprg.mhx to test it. (on larger files it takes *very*
long, due to a built-in debug delay ;))

todo: after pkernel works correctly, we have to reduce the .mhx size.
when the mhx file is small enough we can use the bootrom hack to load it
directly into ram. after that it should be done :)